+++ internal/utils/utils.go
@@ -5,16 +5,16 @@ import (
 	"reflect"
 )
 
-func CalculateStructSize(s any) (int, error) {
+func CalculateStructSize(s any) (uint32, error) {
 	v := reflect.ValueOf(s)
 	if v.Kind() != reflect.Struct {
 		return 0, errors.New("input is not a struct")
 	}
 
-	size := 0
+	size := uint32(0)
 	for i := 0; i < v.NumField(); i++ {
 		field := v.Field(i)
-		fieldSize := int(reflect.TypeOf(field.Interface()).Size())
+		fieldSize := uint32(reflect.TypeOf(field.Interface()).Size())
 		size += fieldSize
 	}
